// src/types/translationKeys.ts
export type SapphireFueKeys =
  | 'meta.title'
  | 'meta.description'
  | 'meta.keywords'
  | 'title'
  | 'description'
  | 'benefits'
  | 'risks'
  | 'preparation'
  | 'recovery';

export type DHIKeys =
  | 'meta.title'
  | 'meta.description'
  | 'meta.keywords'
  | 'title'
  | 'description'
  | 'benefits'
  | 'risks'
  | 'preparation'
  | 'recovery';

export type StemCellKeys =
  | 'meta.title'
  | 'meta.description'
  | 'meta.keywords'
  | 'title'
  | 'description'
  | 'benefits'
  | 'risks'
  | 'preparation'
  | 'recovery';

export type TranslationKeys = SapphireFueKeys | DHIKeys | StemCellKeys;